chore: move lib.rs to tests.rs
